Move Uber and Lyft into the Sandbox part of the code.
