+++ src/ext-test/resources/updater/__files/lyft_authentication.json
