+++ src/ext-test/resources/updater/__files/lyft_trip_estimates.json
