+++ src/ext-test/resources/updater/__files/uber_trip_estimates.json
