+++ src/ext/java/org/opentripplanner/ext/tnc/updater/TransportationNetworkCompanyDataSource.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.updater.transportation_network_company;
+package org.opentripplanner.ext.tnc.updater;
 
 import com.google.common.cache.Cache;
 import com.google.common.cache.CacheBuilder;
