+++ src/ext/java/org/opentripplanner/ext/tnc/updater/lyft/LyftArrivalEstimate.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.updater.transportation_network_company.lyft;
+package org.opentripplanner.ext.tnc.updater.lyft;
 
 public class LyftArrivalEstimate {
     public int eta_seconds;
