+++ src/ext/java/org/opentripplanner/ext/tnc/updater/uber/UberArrivalEstimateResponse.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.updater.transportation_network_company.uber;
+package org.opentripplanner.ext.tnc.updater.uber;
 
 import java.util.List;
